don't need bcrypt and bcryptjs, remove comments, removed duped attrs in userOrg entity, move enums to constants file, fix CreateUserOrganizationDto
