+++ server/package.json
@@ -35,7 +35,6 @@
         "@nestjs/serve-static": "^2.2.2",
         "@nestjs/typeorm": "^7.1.5",
         "@types/cookie-parser": "^1.4.2",
-        "bcrypt": "^5.0.1",
         "bcryptjs": "^2.4.3",
         "config": "^3.3.6",
         "cookie-parser": "^1.4.6",
